tools: fix `crowtty` cargo alias/just recipe (#344)
PR #342 broke the `cargo crowtty` alias and `just crowtty` Just recipe.
This is because we moved the `crowtty` binary out of the workspace's
`default-packages`, so that CI builds don't require `libuv`. Because
`crowtty` is no longer in `default-package`, running it using `cargo
run` now requires passing --pacakge crowtty` as well as `--bin crowtty`,
which the aliases don't do.

This commit fixes that by adding `--package crowtty` to the `cargo
crowtty` alias and the Just recipe. I also documented why crowtty isn't
in `default-packages` in the Cargo.toml.